Resolved: Moved getVenueInfo test helper function to parent class
