+++ notifyme-sdk-backend/notifyme-sdk-backend-ws/src/test/java/ch/ubique/notifyme/sdk/backend/ws/insertmanager/insertfilters/FakeRequestFilterTest.java
@@ -30,33 +30,4 @@ public class FakeRequestFilterTest extends UploadInsertionFilterTest {
   UploadInsertionFilter insertionFilter() {
     return new FakeRequestFilter();
   }
-
-  private UploadVenueInfo getVenueInfo(boolean fake) {
-    LocalDateTime start = LocalDateTime.now();
-    LocalDateTime end = start.plusMinutes(30);
-    final var crypto = cryptoWrapper.getCryptoUtilV3();
-    final var noncesAndNotificationKey =
-        crypto.getNoncesAndNotificationKey(crypto.createNonce(256));
-    byte[] preid =
-        crypto.cryptoHashSHA256(
-            crypto.concatenate(
-                "CN-PREID".getBytes(StandardCharsets.US_ASCII),
-                "payload".getBytes(StandardCharsets.US_ASCII),
-                noncesAndNotificationKey.noncePreId));
-    byte[] timekey =
-        crypto.cryptoHashSHA256(
-            crypto.concatenate(
-                "CN-TIMEKEY".getBytes(StandardCharsets.US_ASCII),
-                crypto.longToBytes(3600L),
-                crypto.longToBytes(start.toInstant(ZoneOffset.UTC).getEpochSecond()),
-                noncesAndNotificationKey.nonceTimekey));
-    return UploadVenueInfo.newBuilder()
-        .setPreId(ByteString.copyFrom(preid))
-        .setTimeKey(ByteString.copyFrom(timekey))
-        .setIntervalStartMs(start.toInstant(ZoneOffset.UTC).toEpochMilli())
-        .setIntervalEndMs(end.toInstant(ZoneOffset.UTC).toEpochMilli())
-        .setNotificationKey(ByteString.copyFrom(noncesAndNotificationKey.notificationKey))
-        .setFake(fake)
-        .build();
-  }
 }
