+++ notifyme-sdk-backend/notifyme-sdk-backend-ws/src/test/java/ch/ubique/notifyme/sdk/backend/ws/insertmanager/insertfilters/UploadInsertionFilterTest.java
@@ -5,6 +5,9 @@ import ch.ubique.notifyme.sdk.backend.ws.insertmanager.OSType;
 import ch.ubique.notifyme.sdk.backend.ws.semver.Version;
 import ch.ubique.notifyme.sdk.backend.ws.util.CryptoWrapper;
 import ch.ubique.notifyme.sdk.backend.ws.util.TokenHelper;
+import com.google.protobuf.ByteString;
+import java.nio.charset.StandardCharsets;
+import org.apache.tomcat.jni.Local;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -79,5 +82,43 @@ public abstract class UploadInsertionFilterTest {
         insertionFilter()
             .filter(now, uploadVenueInfoList, osType, osVersion, appVersion, token)
             .isEmpty());
+  }
+
+  public UploadVenueInfo getVenueInfo(LocalDateTime start, LocalDateTime end) {
+    return getVenueInfo(start, end, false);
+  }
+
+  public UploadVenueInfo getVenueInfo(boolean fake) {
+    LocalDateTime start = LocalDateTime.now();
+    LocalDateTime end = start.plusMinutes(30);
+    return getVenueInfo(start, end, fake);
+  }
+
+  public UploadVenueInfo getVenueInfo(
+      LocalDateTime start, LocalDateTime end, boolean fake) {
+    final var crypto = cryptoWrapper.getCryptoUtilV3();
+    final var noncesAndNotificationKey =
+        crypto.getNoncesAndNotificationKey(crypto.createNonce(256));
+    byte[] preid =
+        crypto.cryptoHashSHA256(
+            crypto.concatenate(
+                "CN-PREID".getBytes(StandardCharsets.US_ASCII),
+                "payload".getBytes(StandardCharsets.US_ASCII),
+                noncesAndNotificationKey.noncePreId));
+    byte[] timekey =
+        crypto.cryptoHashSHA256(
+            crypto.concatenate(
+                "CN-TIMEKEY".getBytes(StandardCharsets.US_ASCII),
+                crypto.longToBytes(3600L),
+                crypto.longToBytes(start.toInstant(ZoneOffset.UTC).getEpochSecond()),
+                noncesAndNotificationKey.nonceTimekey));
+    return UploadVenueInfo.newBuilder()
+        .setPreId(ByteString.copyFrom(preid))
+        .setTimeKey(ByteString.copyFrom(timekey))
+        .setIntervalStartMs(start.toInstant(ZoneOffset.UTC).toEpochMilli())
+        .setIntervalEndMs(end.toInstant(ZoneOffset.UTC).toEpochMilli())
+        .setNotificationKey(ByteString.copyFrom(noncesAndNotificationKey.notificationKey))
+        .setFake(fake)
+        .build();
   }
 }
